requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
